+++ game_dir/chess_engine/pieces/__init__.py
