+++ game_dir/chess_engine/pieces/king.py
@@ -1,6 +1,6 @@
-from game_files.chess_engine.piece import Piece
-from game_files.chess_engine.pieces.rook import Rook
-from game_files.chess_engine.position import Position
+from game_dir.chess_engine.piece import Piece
+from game_dir.chess_engine.pieces.rook import Rook
+from game_dir.chess_engine.position import Position
 
 
 class King(Piece):
