+++ game_dir/chess_engine/stockfish/AUTHORS
