+++ game_dir/chess_engine/stockfish/CITATION.cff
