+++ game_dir/chess_engine/stockfish/CONTRIBUTING.md
