+++ game_dir/utils/images/bB.png
