+++ game_dir/utils/images/bK.png
