+++ game_dir/utils/images/bP.png
