+++ game_dir/utils/images/bQ.png
