+++ game_dir/utils/images/bR.png
