+++ game_dir/utils/images/cross.png
