+++ game_dir/utils/images/crown.png
