+++ game_dir/utils/images/wB.png
