+++ game_dir/utils/images/wK.png
