+++ game_dir/utils/images/wN.png
