+++ game_dir/utils/images/wP.png
