+++ game_dir/utils/images/wQ.png
