+++ game_dir/utils/images/wR.png
