+++ requirements.txt
@@ -1,4 +1,4 @@
 pygame
 pydantic
 stockfish
-chess
+json
